refactor: remove unnecessary `heroic_` prefix from filenames
