+++ src/linux/launchers/heroic/amazon.rs
