+++ src/linux/launchers/heroic/gog.rs
